S02EXX - Sokoban - Offline Development - HSP was having a problem because it was Greedy and Not A*. - HSP was having a problem with Infinity after I changed it to A*

+++ Assets/Scripts/Planning/HeuristicSearchPlannerSGW/HSPHeuristic.cs
@@ -47,7 +47,7 @@ namespace HeuristicSearchPlannerSGW
 
         protected abstract int cost(Expression expression);
 
-        public override double evaluate(State current)
+        public override int evaluate(State current)
         {
             costMap.Clear();
             foreach (Literal literal in literals)
